PhasorNetworks

Documentation for PhasorNetworks.

PhasorNetworks.PhasorSA

PhasorNetworks.match_offsets

PhasorNetworks.match_offsets

PhasorNetworks.phase_to_potential

PhasorNetworks.phase_to_time

PhasorNetworks.potential_to_phase

PhasorNetworks.similarity_outer

PhasorNetworks.PhasorSA — Type

Phasor Self-Attention Module

PhasorNetworks.match_offsets — Method

Delay spike trains as necessary to make the represented phases between them match

PhasorNetworks.match_offsets — Method

Delay the spike trains in a vector as necessary to make their offsets match

PhasorNetworks.phase_to_potential — Method

Convert a static phase to the complex potential of an R&F neuron

PhasorNetworks.phase_to_time — Method

Converts a matrix of phases into a spike train via phase encoding

phasetotrain(phases::AbstractMatrix, spk_args::SpikingArgs, repeats::Int = 1, offset::Real = 0.0)

PhasorNetworks.potential_to_phase — Method

Convert the potential of a neuron at an arbitrary point in time to its phase relative to a reference

PhasorNetworks.similarity_outer — Method

Slicing each array along 'dims', find the similarity between each corresponding slice and reduce along 'reduce_dim'